GoogleMapsBulkGeocoder extraction
