+++ server/lib/python/cartodb_services/cartodb_services/google/__init__.py
@@ -1,1 +1,2 @@
 from geocoder import GoogleMapsGeocoder
+from bulk_geocoder import GoogleMapsBulkGeocoder
